Added support for kernels without autosuspend.

+++ debian/init.d
@@ -4,6 +4,8 @@ DESC="autosuspend"
 NAME="autosuspend-support"
 PATH=/usr/local/sbin:/usr/local/bin:/sbin:/bin:/usr/sbin:/usr/bin
 
+SYS_CTL=/sys/power/autosuspend
+
 [ -f /etc/default/autosuspend ] && . /etc/default/autosuspend
 
 set_timeout() {
@@ -15,22 +17,34 @@ set_timeout() {
 case "$1" in
 	start)
 		echo -n "Starting $DESC: "
-        set_timeout
-		echo 1 > /sys/power/autosuspend
-		echo "$NAME."
+        if [ -f $SYS_CTL ]; then
+            set_timeout
+		    echo 1 > $SYS_CTL
+		    echo ok.
+        else
+            echo "not supported by kernel."
+        fi
 		;;
 	stop)
 		echo -n "Stopping $DESC: "
-		echo 0 > /sys/power/autosuspend
-		echo "$NAME."
+        if [ -f $SYS_CTL ]; then
+		    echo 0 > /sys/power/autosuspend
+		    echo ok.
+        else
+            echo "not supported by kernel."
+        fi
 		;;
     force-reload)
         echo -n "Reloading $DESC: "
-        set_timeout
-        echo "$NAME."
+        if [ -f $SYS_CTL ]; then
+            set_timeout
+            echo ok.
+        else
+            echo "not supported by kernel."
+        fi
         ;;
 	*)
-		echo "Usage: /etc/init.d/$NAME {start|stop|force-reload}" >&2
+		echo "Usage: /etc/init.d/autosuspend-support {start|stop|force-reload}" >&2
 		exit 1
 	    ;;
 esac
